+++ src/main/java/com/shanebeestudios/skbee/api/bound/BoundConfig.java
@@ -1,10 +1,10 @@
-package com.shanebeestudios.skbee.elements.bound.config;
+package com.shanebeestudios.skbee.api.bound;
 
 import ch.njol.skript.Skript;
 import com.shanebeestudios.skbee.SkBee;
 import com.shanebeestudios.skbee.api.util.Util;
 import com.shanebeestudios.skbee.api.util.WorldUtils;
-import com.shanebeestudios.skbee.elements.bound.objects.Bound;
+import com.shanebeestudios.skbee.api.bound.Bound;
 import org.bukkit.World;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.FileConfiguration;
