+++ src/main/java/com/shanebeestudios/skbee/api/listener/BoundBorderListener.java
@@ -3,8 +3,8 @@ package com.shanebeestudios.skbee.api.listener;
 import com.shanebeestudios.skbee.SkBee;
 import com.shanebeestudios.skbee.api.event.EnterBoundEvent;
 import com.shanebeestudios.skbee.api.event.ExitBoundEvent;
-import com.shanebeestudios.skbee.elements.bound.config.BoundConfig;
-import com.shanebeestudios.skbee.elements.bound.objects.Bound;
+import com.shanebeestudios.skbee.api.bound.BoundConfig;
+import com.shanebeestudios.skbee.api.bound.Bound;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
