+++ src/main/java/com/shanebeestudios/skbee/elements/bound/expressions/ExprBoundAtLocation.java
@@ -13,7 +13,7 @@ import ch.njol.util.Kleenean;
 import com.shanebeestudios.skbee.SkBee;
 import org.bukkit.Location;
 import org.bukkit.event.Event;
-import com.shanebeestudios.skbee.elements.bound.objects.Bound;
+import com.shanebeestudios.skbee.api.bound.Bound;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
